Add new net. Fix how the position was flipped/mirrored for the black perspective in the NNUE.
Instead of 'rotating' the position, the position is now vertically mirrored.
This ensures that the king and queen have the same feature index in the start position for both perspectives.
The new net in this commit, is also improved due to fixes in previous commits,
where the the target WDL was not correct for the black perspective